extract common roles for bootstrapper and miner
